Improved ProjectGenerator
